Add *nix tests to validate mutex works as well
