+++ source/Calamari.Tests/Fixtures/Integration/Process/Semaphores/NixSystemSemaphoreFixture.cs
@@ -1,0 +1,11 @@
+using Calamari.Testing.Helpers;
+using NUnit.Framework;
+
+namespace Calamari.Tests.Fixtures.Integration.Process.Semaphores
+{
+    [TestFixture]
+    [Category(TestCategory.CompatibleOS.OnlyNixOrMac)]
+    public class NixSystemSemaphoreFixture : SemaphoreFixtureBase
+    {
+    }
+}
