+++ source/Calamari.Tests/Fixtures/Integration/Process/Semaphores/WindowsSystemSemaphoreFixture.cs
@@ -1,0 +1,10 @@
+using Calamari.Testing.Helpers;
+using NUnit.Framework;
+
+namespace Calamari.Tests.Fixtures.Integration.Process.Semaphores
+{
+    [TestFixture]
+    [Category(TestCategory.CompatibleOS.OnlyWindows)]
+    public class WindowsSystemSemaphoreFixture : SemaphoreFixtureBase
+    { }
+}
